+++ src/polyfill/index.js
@@ -1,0 +1,6 @@
+// https://github.com/melonjs/melonJS/issues/1092
+import "core-js/proposals/global-this";
+
+// "built-in" polyfills
+import "./console.js";
+import "./requestAnimationFrame.js";