+++ src/polyfill/requestAnimationFrame.js
@@ -6,23 +6,23 @@ var x;
 
 // standardized functions
 // https://developer.mozilla.org/fr/docs/Web/API/Window/requestAnimationFrame
-var requestAnimationFrame = window.requestAnimationFrame;
-var cancelAnimationFrame = window.cancelAnimationFrame;
+var requestAnimationFrame = globalThis.requestAnimationFrame;
+var cancelAnimationFrame = globalThis.cancelAnimationFrame;
 
 // get prefixed rAF and cAF is standard one not supported
 for (x = 0; x < vendors.length && !requestAnimationFrame; ++x) {
-    requestAnimationFrame = window[vendors[x] + "RequestAnimationFrame"];
+    requestAnimationFrame = globalThis[vendors[x] + "RequestAnimationFrame"];
 }
 for (x = 0; x < vendors.length && !cancelAnimationFrame; ++x) {
-    cancelAnimationFrame = window[vendors[x] + "CancelAnimationFrame"] ||
-                           window[vendors[x] + "CancelRequestAnimationFrame"];
+    cancelAnimationFrame = globalThis[vendors[x] + "CancelAnimationFrame"] ||
+                           globalThis[vendors[x] + "CancelRequestAnimationFrame"];
 }
 
 if (!requestAnimationFrame || !cancelAnimationFrame) {
     requestAnimationFrame = function (callback) {
-        var currTime = window.performance.now();
+        var currTime = globalThis.performance.now();
         var timeToCall = Math.max(0, (1000 / timer.maxfps) - (currTime - lastTime));
-        var id = window.setTimeout(function () {
+        var id = globalThis.setTimeout(function () {
             callback(currTime + timeToCall);
         }, timeToCall);
         lastTime = currTime + timeToCall;
@@ -30,10 +30,10 @@ if (!requestAnimationFrame || !cancelAnimationFrame) {
     };
 
     cancelAnimationFrame = function (id) {
-        window.clearTimeout(id);
+        globalThis.clearTimeout(id);
     };
 
     // put back in global namespace
-    window.requestAnimationFrame = requestAnimationFrame;
-    window.cancelAnimationFrame = cancelAnimationFrame;
+    globalThis.requestAnimationFrame = requestAnimationFrame;
+    globalThis.cancelAnimationFrame = cancelAnimationFrame;
 }